Rename Home component to User and update routing accordingly
