--- example/src/components/Home/index.ts
@@ -1,1 +1,0 @@
-export * from './Home';
